requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
bitsandbytes==0.42.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
clip==1.0
contourpy==1.3.0
cycler==0.12.1
datasets==4.4.1
dill==0.4.0
distro==1.9.0
exceptiongroup==1.3.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
kagglehub==0.3.13
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.2.1
numpy==2.0.2
openai==2.9.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==0.11.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.3
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
wcwidth==0.2.14
Werkzeug==3.1.4
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
